backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
mysql-connector-python==8.3.0
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
SQLAlchemy==2.0.25
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2